Made changes as per the requirement and added the functionality of report generation

+++ Pharmacy/Models/PurchaseInvoice.cs
@@ -25,7 +25,7 @@ namespace Pharmacy.Models
         public string InvoiceId { get; set; }
         public Nullable<int> supplierId { get; set; }
         public Nullable<System.DateTime> DateOfPurchase { get; set; }
-        public Nullable<decimal> TotalPurchaseAmount { get; set; }
+        public Nullable<double> TotalPurchaseAmount { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<purchase> purchases { get; set; }
