+++ Pharmacy/ViewModels/SupplierLedgerViewModel.cs
@@ -6,7 +6,7 @@ namespace Pharmacy.ViewModels
     {
         public int Id { get; set; }
         public int? SupplierId { get; set; }
-        public int? BalanceAmount { get; set; }
+        public double? BalanceAmount { get; set; }
         public int? CreditAmount { get; set; }
         public int? DebitAmunt { get; set; }
         public string Particulars { get; set; }
