+++ Pharmacy/ViewModels/ViewSales.cs
@@ -7,7 +7,7 @@ namespace Pharmacy.ViewModels
         public int Id { get; set; }
         public string InvoiceId { get; set; }
         public DateTime? DateOfPurchase { get; set; }
-        public decimal? PurchaseAmount { get; set; }
+        public double? PurchaseAmount { get; set; }
         public int? CustomerId { get; set; }
     }
 }
